Reworked system to change the plots properties on runtime ( communication between OGLPlotRendererWidget_C and PlotModel_C)
The plot model deos not execute commands ( which modify the plot objects), but instead it sends events, to which the OGLPlotRendererWidget_C listens.
This guarantees, that no Gl-Call is issued at the wrong time, but in a defined order.

in short: Changes to the plots are now applied from the OGLPlotRendererWidget_C.
The PlotModel sends the events (triggered by user) to which the PlotRendererWidget listens.
These events are stored inside an event-queue(_request_buffer) in the OGLPlotRendererWidget_C
and processed before each gl Draw() call (paintGL-loop)

WIP